src/dataviewer/backend/pyproject.toml
[project]
name = "lerobot-annotation-api"
version = "0.1.0"
description = "API for LeRobot episode annotation system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
azure = [
    "azure-storage-blob>=12.24.0",
    "azure-identity>=1.19.0",
]
analysis = [
    "numpy>=2.2.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.6.0",
    "pandas>=2.2.0",
    "pyarrow>=18.1.0",
]
huggingface = [
    "huggingface-hub>=0.27.0",
]
hdf5 = [
    "h5py>=3.12.0",
]
export = [
    "h5py>=3.12.0",
    "numpy>=2.2.0",
    "Pillow>=10.0.0",
]
yolo = [
    "ultralytics>=8.3.0",
    "numpy>=2.2.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
# B008: Depends() in FastAPI is a valid pattern
# B905: zip() without strict= is fine for most use cases
# B904: Exception chaining - we pass cause to constructor
ignore = ["B008", "B905", "B904"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
